pal: Fix Device::GetHdrMetaData for Linux/amdgpu.
The EDID parsing contains bugs, as confirmed by comparing the
implementation against the implementation of the edid-decode
tool (https://git.linuxtv.org/edid-decode.git/tree/parse-base-block.cpp)
and relevant EDID specs. This also shows when comparing the
results from amdvlk on Linux against the results of the AMD
Vulkan driver on Windows-10, where the Windows driver reports
correct chromaticity coordinates for the blue primary and the
white point, consistent with the output of edid-decode.

This fixes the bugs and makes results consistent with edid-decode
and the Vulkan driver for Windows-10.

When we are at it, fix a missing assignment of
connectorSupportHdr = true, which made Device::GetHdrMetaData()
return failure even though query of HDR display capabilities was
successfull. Not that it matters, as the Screen class does not
actually checkt the return value of Device::GetHdrMetaData, but
for completeness.

Signed-off-by: Mario Kleiner <mario.kleiner.de@gmail.com>