pal: Only set 'max bpc' = 10 in HDR mode if it is currently lower.
For HDR-10 we need a 'max bpc' limit on the DRM connector of at
least 10 bpc, but the user or 3rd party code may have already
set a limit of > 10 bpc. Do not interfere with the users choice,
as long as it is at least the minimum required 10 bpc for HDR.

This is, e.g., useful with the new fp16 swapchain format, where
the framebuffer can provide a net precision of more than 10 bpc,
so we don't want to truncate that back to only 10 bpc.

Signed-off-by: Mario Kleiner <mario.kleiner.de@gmail.com>