Restore full 24 or 16 bit depth buffer precision on gfx6/9
The lowered precision of 22 or 15 bits reduces overall visual
quality of scenes with shadow maps, even if rounding may be
more accurate with fewer bits. This change restores the full
24 or 16 bits for UNORM depth buffers.

Tested by running examples and confirming the results look more natural
and match previous renders. See
https://github.com/SaschaWillems/Vulkan/blob/master/examples/shadowmapping/shadowmapping.cpp
for an example scene that has reduced quality with a 15 bit depth buffer. This example uses a depth bias of 1.25f, possibly to avoid the rounding
issue.

Fixes #49