Fix custom operator new for clang with noexcept
clang ignores gcc's -fcheck-new flag, which means that the few custom
operator new implementations in AMDVLK did not handle OOM conditions
properly and invoked constructors with null this pointers. But
fortunately, all of theose operators are custom user operators, by
virtue of using a custom third argument. Consequently, they are allowed
to be noexcept, which will force the compiler to emit a null check. This
is a better fix than using -fcheck-new and will work with any standard
C++ compiler.